+++ src/Model/SettingValue.php
@@ -68,7 +68,7 @@ class SettingValue implements SettingValueInterface
     /**
      * @return string
      */
-    public function getSettingName(): string
+    public function getName(): string
     {
         return $this->settingName;
     }